% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac1.R
\name{ac1}
\alias{ac1}
\title{AC1
Calculate first-order autocorrelation}
\usage{
ac1(x, ...)
}
\arguments{
\item{x}{numeric vector of values of a regular (equally-spaced) time series}

\item{...}{unused, here for compatibility with functions that might, by default, pass arguments unused by this function}
}
\value{
AR(1) coefficient
}
\description{
AC1
Calculate first-order autocorrelation
}

