% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac1.R
\name{ac_sub}
\alias{ac_sub}
\title{Subset AC1}
\usage{
ac_sub(x, n, phase, ...)
}
\arguments{
\item{x}{numeric vector, time series}

\item{n}{integer, subset to every n-th observation (every n-th pair of observations)}

\item{phase}{integer, initial observation (pair) to use when subsetting, passed to \code{\link{sub_samp}}}

\item{...}{unused, here for compatability with other functions that might pass arguments by default

To do a subset for autocorrelation, the key is retain the adjacency of observations. For \code{n=2}, for example, we couldn't use every other observation in \code{x}, as this would fundamentally change the time scale of the calculation. Thus, we still want to calculate the correlation between adjacent samples. So in this example, we would use every other adjacent pair. E.g., samples 1-2, (skip 2-3), 3-4, (skip 4-5), ... and so on.}
}
\value{
AR1 coefficient
}
\description{
Calculate first-order autocorrelation from a subset
}

