% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_roll.R
\name{acf_cor}
\alias{acf_cor}
\title{ACF Correlation}
\usage{
acf_cor(x, ...)
}
\arguments{
\item{x}{time series}

\item{...}{paramters to pass to \code{stats::acf}}
}
\value{
numeric vector of correlations
}
\description{
Helper function to calculate autocorrelation across many time series and then just return the correlations
}

