% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_roll.R
\name{acf_roll}
\alias{acf_roll}
\title{Rolling Window ACF}
\usage{
acf_roll(x, width = 28 * 24, by = 12, lag.max = width/14,
  DETREND = FALSE, mp = 1, mf = 2, mi = 1, ...)
}
\arguments{
\item{x}{the time series}

\item{width}{width of rolling window}

\item{by}{increment window by this many time steps (n observations)}

\item{lag.max}{integer indicating the maximum lag at which to calculate the ACF (as in code{stats::acf})}

\item{DETREND}{whether or not to perform AICc-selected detrending; see \code{\link{detrendR}}}

\item{mp, mf, mi}{maximum order of polynomial, Fourier, and interaction}

\item{...}{unused}
}
\description{
Performs rolling window autocorrelation calculation across many time scales
}
\seealso{
\code{stats::ts} \code{forecast::fourier} \code{\link{trend_xreg}} \code{\link{interaction_xreg}}
}

