% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_ts.R
\name{agg_ts}
\alias{agg_ts}
\title{Agg TS}
\usage{
agg_ts(x, y, width = 288, na.rm = TRUE, FUN = mean)
}
\arguments{
\item{x}{time index vector}

\item{y}{value vector (numeric, or a \code{\link{ts}} object)}

\item{width}{integer indicating number of observations to aggregate at a time}

\item{na.rm}{logical, remove NA's? passed to \code{FUN}}

\item{FUN}{aggregation function}
}
\value{
data.table with 3 columns: x, y, and N. N is number of observation per aggregated value (number of observations for input to \code{FUN}). If the input y was a time series, so will the output
}
\description{
Aggregate Time Series
}

