% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arEigs.R
\name{arEigs}
\alias{arEigs}
\title{Eigenvalues}
\usage{
arEigs(b)
}
\arguments{
\item{b}{numeric vector of AR coefficients}
}
\value{
eigenvalues
}
\description{
Eigenvalues
}
\seealso{
\code{\link{diagExtend}}, \code{\link{ar}}
}

