% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arP.R
\name{arP}
\alias{arP}
\title{AR(p)}
\usage{
arP(x, method = c("ar", "auto"), stepwise = TRUE)
}
\arguments{
\item{x}{time series or numeric vector}

\item{method}{either 'ar' or 'auto'; 'ar' uses \code{ar} in the 'stats' package, 'auto' uses the \code{auto.arima} in the 'forecast' package.}

\item{stepwise}{logical, if method is "auto", indicates whether to perform stepwise model selection (fast) or search all models. As in \code{auto.arima}}
}
\value{
ar coefficients
}
\description{
Fit an AR(p) Model
}

