% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvarss.R
\name{as.tvarss}
\alias{as.tvarss}
\title{Convert to Class 'tvarss'}
\usage{
as.tvarss(x)
}
\arguments{
\item{x}{an object of class 'rjags'}
}
\value{
an object of class 'tvarss'
}
\description{
Converts an object of class 'rjags' to class 'tvarss', which is just s subset of 'rjags'.
}

