% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-layers.R
\name{calendar_breaks}
\alias{calendar_breaks}
\title{Thinned discrete breaks that keep the end values}
\usage{
calendar_breaks(n = 6)
}
\arguments{
\item{n}{Approximate number of labels (>= 2; default 6).}
}
\value{
A function of the level vector, for \verb{breaks =}.
}
\description{
A breaks \emph{function} for discrete timeslice axes: picks about \code{n}
evenly spaced labels and ALWAYS includes the first and last -- dense
vocabularies (\code{d001..d365}, \code{h00..h23}) stop overlapping without
losing the axis range. Pass it to \code{ggplot2::scale_x_discrete()} /
\code{scale_y_discrete()}:
}
\details{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{ggplot(x) +
  geom_calendar(calendar = cal, datetime = "t", z = "v") +
  scale_x_discrete(breaks = calendar_breaks()) +
  theme_calendar()
}\if{html}{\out{</div>}}
}
