% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapLegend.R
\name{colorBar}
\alias{colorBar}
\title{Segments for Map Legend}
\usage{
colorBar(bar1, bar2, cols, n = 256)
}
\arguments{
\item{bar1, bar2}{length 4 vector containing x0,y0,x1,y1 elements (see \code{\link{segments}})}

\item{cols}{colors used in heat map figure}

\item{n}{number of segments that will form color bar}
}
\description{
Create the segments that will form the color bar in a map legend
}
\seealso{
\code{\link{mapLegend}}
}

