% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLakeP.R
\name{d2Xdt}
\alias{d2Xdt}
\title{Derivative of dX/dt}
\usage{
d2Xdt(X, pars)
}
\arguments{
\item{X}{numeric, value of water P}

\item{pars}{optional named vector of parameters; parameters not supplied here will be taken from defaults for \code{\link{modelDeterministicXM}}}
}
\value{
numeric vector
}
\description{
Function that calculates the derivative of dX/dt when dM/dt is 0
}

