% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend.R
\name{detrend}
\alias{detrend}
\title{Detrend
Detrend time series}
\usage{
detrend(x, time = 1:length(x))
}
\arguments{
\item{x}{vector time series}

\item{time}{vector of time steps}
}
\value{
detrended x
}
\description{
Detrend
Detrend time series
}

