% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend.R
\name{detrendR}
\alias{detrendR}
\title{DetrendR}
\usage{
detrendR(x, max_poly = 6, max_fourier = 6, max_interaction = 3)
}
\arguments{
\item{x}{time series to be detrended}

\item{max_poly}{integer indicating the order of the polynomial trend}

\item{max_fourier}{integer indicating the order of the Fourier series}

\item{max_interaction}{integer indicating the polynomial order that Fourier series should interact with (see \code{\link{interaction_xreg}})}
}
\value{
detrended time series
}
\description{
Detrending of time series using polynomial trend, Fourier series, and their interaction; includes model selection
}
\details{
model to be used for detrending is fit via OLS, the best model is selected by AICc. See \code{\link{interaction_xreg}} for the construction of the interaction matrix. Not all model combinations are considered just all combinations of the orders of polynomial and Fourier series and interactions of the Fourier series with main-effect polynomials; the order of the interaction will never be greater than either of the orders of the polynomial or Fourier. Missing values (NA's) are first linearly interpolated.
}
\seealso{
\code{\link{detrend}}\code{stats::ts} \code{forecast::fourier} \code{\link{trend_xreg}} \code{\link{interaction_xreg}} \code{\link{fill_na}}
}

