% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLakeP.R
\name{dX_dt_ofXI}
\alias{dX_dt_ofXI}
\title{Change in X with respect to time as a function of X when dM/dt is 0}
\usage{
dX_dt_ofXI(X, I, pars)
}
\arguments{
\item{X}{numeric, water P}

\item{I}{numeric, P input to lake}

\item{pars}{optional named vector of parameters; parameters not supplied (or if pars is not specificied at all) will default to defaults in \code{\link{modelDeterministicXM}}}
}
\value{
a numeric value indicating dX/dt
}
\description{
Take original model, set dM/dt to 0 and solve for M, plug into dX/dt, which gives us dX/dt as a function of X when dM/dt is 0.  Calculates the change in water P per unit time as a function of water P and parameters.
}
\seealso{
\code{\link{modelDeterministicXM}}
}

