% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na.R
\name{fill_na}
\alias{fill_na}
\title{Fill-in NA's}
\usage{
fill_na(x)
}
\arguments{
\item{x}{vector with NA's that is to be interpolated}
}
\value{
a numeric vector or a \code{ts} object
}
\description{
Fill NA valus with linearly interpolated values
}
\details{
If starting or ended values are NA, repeats nearest non-NA value. If \code{x} is a time series (object of class \code{ts}), the output is also a \code{ts}.
}
\seealso{
\code{\link{approx}} \code{\link{ts}}
}

