% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{get_calendar_rule}
\alias{get_calendar_rule}
\title{Look up a registered rule}
\usage{
get_calendar_rule(param)
}
\arguments{
\item{param}{Name of the value column.}
}
\value{
A list with element \code{rule}, or \code{NULL} if \code{param} has not been
registered.
}
\description{
Look up a registered rule
}
\examples{
register_calendar_rule("demand", "sum")
get_calendar_rule("demand")
get_calendar_rule("not_registered")
}
