% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLakeP.R
\name{getEigs_df}
\alias{getEigs_df}
\title{Get Eigenvalues for Eutrophication Model (from data frame of states)
Given the rate of change of state variables (water and mud P) and the input P, returns the eigenvalues}
\usage{
getEigs_df(rootGrid_I)
}
\arguments{
\item{rootGrid_I}{a matrix with columns named for X, M, I, as in \code{\link{modelDeterministicXM}}; these columns are the equilibria (roots) of the state variables, and I, the P input parameter.}
}
\value{
data.frame of eigenvalues
}
\description{
Get Eigenvalues for Eutrophication Model (from data frame of states)
Given the rate of change of state variables (water and mud P) and the input P, returns the eigenvalues
}
\note{
the "Grid" name just comes from legacy coding where I was trying a "grid" (as in, e.g., \code{expand.grid}) of X, M, and I values as initial values for which I was calculating equilibria.
}
\seealso{
\code{\link{getEigs}}
}

