% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLakeP.R
\name{getInit}
\alias{getInit}
\title{Get Initial Values for Eutrohpication Model
Gets a grid/ mesh of values for eventual input to model}
\usage{
getInit(gridN = 100, Irange = c(0.25, 1.75), Xrange = c(0.05, 8),
  Mrange = c(400, 400), In = gridN, Xn = gridN, Mn = gridN)
}
\arguments{
\item{gridN}{number of values for I and X}

\item{Irange}{range of values for I, input rate of P to lake}

\item{Xrange}{range of values for X, water phosphorus}

\item{Mrange}{range of values for M, mud P}

\item{In}{number of values for I; default is set to gridN}

\item{Xn}{number of values for X; default is set to gridN}

\item{Mn}{number of values for M; default is set to gridN}
}
\value{
a data.frame of initial values
}
\description{
Get Initial Values for Eutrohpication Model
Gets a grid/ mesh of values for eventual input to model
}

