% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLakeP.R
\name{getRoot}
\alias{getRoot}
\title{Get the Root of the Eutrophication Model
Finds the roots of the eutrophication model given starting values (water, mud P) and parameter (P loading)}
\usage{
getRoot(x, maxiter = 1000, ...)
}
\arguments{
\item{x}{a vector of length 3 with names X (water P), M (mud P), and I (P loading)}

\item{maxiter}{maximum number of iterations to use when finding root}

\item{...}{additional arguments to pass to \code{\link{modelDeterministicXM}}}
}
\value{
numeric vector with roots
}
\description{
Get the Root of the Eutrophication Model
Finds the roots of the eutrophication model given starting values (water, mud P) and parameter (P loading)
}

