% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_xreg.R
\name{interaction_xreg}
\alias{interaction_xreg}
\title{Interaction (X) Regressors}
\usage{
interaction_xreg(p, f, i = NULL)
}
\arguments{
\item{p}{matrix of polynomial trends, such as from \code{trend_xreg}}

\item{f}{matrix of fourier xregs, such as from \code{forecast::fourier}}

\item{i}{optional integer indicating the number of columns of p to be used in interaction; if NULL (default), interactions are with all column of p, and output is a list. If \code{i} is not NULL, output is a matrix.}
}
\value{
a matrix with columns that are interactions between p and f, or a list where each element is all columns of f interacting with one of the columns of p, and each element of that list uses a different column of p
}
\description{
Given two matrices of predictors p and f, get matrix of all columns of f interacting with each column of p
}
\examples{
freq <- 24
nDays <- 5
ampSlope <- 0.25
ts_time <- 1:(freq*nDays)
sineVec <- sqrt(ts_time)*ampSlope*sin(ts_time*(1/freq)*2*pi)
polyVec <- c(trend_xreg(2, ts_time)\%*\%c(1, 0.75*3.5, -0.78*3))
determVec <- sineVec + polyVec
noiseVec <- rnorm(freq*nDays, sd=0.25)
tsVec <- ts(determVec+noiseVec, frequency=freq)
plot(tsVec, lwd=2)

test_p <- trend_xreg(2, tsVec)
test_f <- forecast::fourier(tsVec, K=2)
fit_mod0 <- lm(tsVec~test_p+test_f)
lines(ts(fitted(fit_mod0), freq=freq), col='blue')

test_i_NULL <- interaction_xreg(p=test_p, f=test_f, i=NULL)
str(test_i_NULL)

test_i_1 <- interaction_xreg(p=test_p, f=test_f, i=1)
str(test_i_1)
fit_mod <- lm(tsVec~test_p+test_f+test_i_1)
lines(ts(fitted(fit_mod), freq=freq), col='red')
}
\seealso{
\code{\link{trend_xreg}} \code{forecast::fourier}
}

