% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_name.R
\name{interval_name}
\alias{interval_name}
\title{Interval Name}
\usage{
interval_name(x, minPerSample = 5)
}
\arguments{
\item{x}{integer, the number of time steps}

\item{minPerSample}{time elapsed between samples, in minutes}
}
\description{
Converts a time interval (integer number of observations) into words, assuming 5 minute data
}
\examples{
interval_name(5)
interval_name(13)
interval_name(288*1.25)

}

