% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{list_calendar_rules}
\alias{list_calendar_rules}
\title{List registered rules}
\usage{
list_calendar_rules()
}
\value{
A \code{data.frame} with columns \code{param} and \code{rule}.
}
\description{
List registered rules
}
\examples{
register_calendar_rule("invcost", "weighted_mean")
list_calendar_rules()
}
