% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapLegend.R
\name{mapLegend}
\alias{mapLegend}
\title{Add a color scale legend to a heat map}
\usage{
mapLegend(x = 0.9, y = 0.2, w = 0.05, h = 0.25, zlim, cols,
  horiz = FALSE, axSide, adj, offset, lab.cex = 1, lab.sig = 2, ...)
}
\arguments{
\item{x, y}{the locations of the center of the legend, as a proportion of plot area (between 0 and 1)}

\item{w, h}{the width and height of the legend, as a proportion of plot area}

\item{zlim}{lower and upper limit of values indicated by colors}

\item{cols}{colors used in the original plot; e.g., created by \code{trawlDiversity::zCol}. Note that it is assumed that color and the value change linearly}

\item{horiz}{Logical, if FALSE (default), the colors in the legend change long the vertical axis}

\item{axSide}{the side of the color bar on which the axis (line, ticks, labels) should be placed; if missing, reasonable default is chosen}

\item{adj}{length 1 or 2 numeric vector to be passed to \code{\link{text}} that will affect the position of the tick marks; if missing, reasonable default is chosen}

\item{offset}{length 1 numeric vector to be passed to \code{\link{text}} that will affect the position of the axis labels; if missing, reasonable default is chosen}

\item{lab.cex}{numeric value to be passed to \code{cex} argument of \code{\link{text}} that will affect the size of the axis labels}

\item{lab.sig}{integer, the number of significant digits to which the axis labels should be rounded; see \code{\link{signif}}}

\item{...}{arguments passed to \code{\link{par}}}
}
\description{
Adds a bar that associates numeric values with colors.  Useful for ay plot that uses a linear color gradient to indicate a numeric value, and for which there are too many colors/ values to specify each individually.
}
\examples{
data(volcano)
v <- volcano
v[v<=quantile(v,0.2)] <- NA # small values to NA (will plot white)
v_cols <- c("blue","white","red") # zCol(256, 1:256)
image(v, col=v_cols)
zl <- range(v, na.rm=TRUE)
mapLegend(x=0.9, y=0.85, zlim=zl, cols=v_cols)
mapLegend(x=0.975, y=0.075, zlim=zl, cols=v_cols, w=0.025, h=0.125)
mapLegend(x=0.73, y=0.975, zlim=zl, cols=v_cols, w=0.125, h=0.025, horiz=TRUE)
mapLegend(x=0.93, y=0.25, zlim=zl, cols=v_cols, w=0.1, h=0.025, horiz=TRUE)

}

