% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLakeP.R
\name{mDXM}
\alias{mDXM}
\title{Wrapper for model when finding stability properties}
\usage{
mDXM(t, y, parameters, ...)
}
\arguments{
\item{t}{time, for compatibility with \code{phaseR} and \code{deSolve} and \code{rootSolve}}

\item{y}{state variables as in \code{\link{modelDeterministicXM}}}

\item{parameters}{same as \code{pars} in \code{\link{modelDeterministicXM}}; had to be called 'parameters' for compatibility with \code{phaseR}}

\item{...}{other arguments to be passed to \code{\link{modelDeterministicXM}}}
}
\value{
rate of change of state variables
}
\description{
Wrapper for model finding stability properties using \code{phaseR} package
}

