% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLakeP.R
\name{mDXM_jac}
\alias{mDXM_jac}
\title{Jacobian matrix for deterministic eutrophication model
Jacboian for the 2D eutrophication model involving water, mud, and input phosphorus}
\usage{
mDXM_jac(t, state, pars)
}
\arguments{
\item{t}{time, for compatability with rootSolve::jacobian.full, but not used otherwise}

\item{state}{state variables as in \code{\link{modelDeterministicXM}}}

\item{pars}{paramters as in \code{\link{modelDeterministicXM}}}
}
\value{
Jacobian matrix
}
\description{
Jacobian matrix for deterministic eutrophication model
Jacboian for the 2D eutrophication model involving water, mud, and input phosphorus
}

