% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLakeP.R
\name{modelDeterministic}
\alias{modelDeterministic}
\title{Eutrophication Bifurcation
Model of lake eutrophication via elevated phosphorus loading. When as phosphorus loading increases, a fold bifurcation is reached, and the system switches into a eutrophic state from a oligotrophic state.}
\usage{
modelDeterministic(state, pars = c(C = 0.00115), F = 14.6, b = 0.001,
  h = 0.15, m = 2.4, r = 0.019, q = 8, s = 0.7, sigma = 0.01)
}
\arguments{
\item{state}{state variables, a vector of length 3 with names "X", "M", and "U"}

\item{pars}{a named vector of parameters; currently only option is "C", a coefficient controlling the fraction of soil P that washes into the lake (units of 'per year'); set up this way so as to be compatible with functions in package "rootSolve" (C is a control parameter affecting water quality).}

\item{F}{input rate of phosporus to soil}

\item{b}{permanent burial rate of phosphorus in sediments}

\item{h}{hydrologic loss (outflow)}

\item{m}{in the recycling function, the value of X at which recycling is half the maximum rate}

\item{r}{recycling coefficient}

\item{q}{in the recycling function, the exponent q determines the slope of R(X) near m}

\item{s}{sedimentation loss}

\item{sigma}{standard deviation of recycling noise}

\item{C}{coefficient for transfer of soil phosphorus to the lake}
}
\value{
change of state variables
}
\description{
Eutrophication Bifurcation
Model of lake eutrophication via elevated phosphorus loading. When as phosphorus loading increases, a fold bifurcation is reached, and the system switches into a eutrophic state from a oligotrophic state.
}

