% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLakeP.R
\name{modelDeterministicXM}
\alias{modelDeterministicXM}
\title{Deterministic Eutrophication Model
A 2D Eutrophication model with state variables for water P (X), mud P (M), and P loading as bifurcation parameter (I)}
\usage{
modelDeterministicXM(state, pars = c(I = 0.00115 * 750), b = 0.001,
  h = 0.252, m = 2.4, r = 0.019, q = 10, s = 0.748)
}
\arguments{
\item{state}{vector of length 2 with names X and M (water and mud P, respectively)}

\item{pars}{named vector of length 1 with name I (P loading)}

\item{b}{P burial coefficient}

\item{h}{outflow coefficient}

\item{m}{the value of water P when recycling from mud to water is half of its maximum value}

\item{r}{a scaling term for recycling between mud and water}

\item{q}{the steepness of the recycling function at \code{m}}

\item{s}{sedimentation rate (turns water P into mud P)}
}
\value{
named vector (X, M) indicating rate of change for water and mud P
}
\description{
Deterministic Eutrophication Model
A 2D Eutrophication model with state variables for water P (X), mud P (M), and P loading as bifurcation parameter (I)
}
\references{
SR Carpenter and WA Brock 2006 Rising variance: a leading indicator of ecological transition. Ecology Letters 9: 311-318
}

