% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasePortrait.R
\name{phasePortrait}
\alias{phasePortrait}
\title{Phase Portrait}
\usage{
phasePortrait(I, pars, func = mDXM, x.lim = c(0, 15), y.lim = c(0, 1000),
  add = FALSE, nFlow = 20, nNull = 300, t.end = 50,
  inits = matrix(c(0.5, 800, 0.1, 950, 3, 700, 1.5, 600, 0.5, 100, 2, 150, 5,
  75, 4.5, 115, 1, 300, 2, 400, 4, 500, 10, 300, 7, 200), byrow = TRUE, ncol =
  2), addLeg = FALSE, legPos = "topright")
}
\arguments{
\item{I}{numeric scalar for P input}

\item{pars}{optional named numeric vector for additional parameters to be passed to \code{func}}

\item{func, }{function that returns local derivatives, defaults to \code{\link{mDXM}}, which is just a formatting wrapper for \code{\link{modelDeterministicXM}}; currently only works for default due to use of \code{\link{getRoot_df}} in \code{\link{stabClass}}}

\item{x.lim, y.lim}{minimum and maximum values for each of the two states variables to be used when plotting and searching for equilibria}

\item{add}{local, if FALSE (default) makes a new plot}

\item{nFlow}{square root of the number of arrows to draw in the vector field}

\item{nNull}{numeric scalar, number of points used in drawing the nullclines}

\item{t.end}{number of time steps (in model time units) to use when simulating the trajectories}

\item{inits}{matrix of initial values to use for trajectories}

\item{addLeg}{if TRUE, adds a legend}

\item{legPos}{character position of legend; see \code{graphics::legend}}
}
\value{
a data.frame with stability information; see \code{\link{stabClass}}
}
\description{
Plot a phase portrait of a dynamical system (defaults to lake eutrophication model) that includes the vector field, nullclines, example trajectories, and the identification and classification of equilibria.
}
\details{
This function is largely a wrapper for \code{phaseR::flowField}, \code{phaseR::trajectory}, \code{phaseR::nullclines}, and \code{phaseR::stability}. This function does not call \code{phaseR::stability} directly, and instead uses \code{\link{stabClass}}.
}
\examples{
shortInits <- eval(formals(phasePortrait)$inits)[1:3,]
pars <- c(q=8,s=0.7,h=0.15, b=0.001)
(phasePortrait(
	I=0.75, pars=pars,
	inits=shortInits, nFlow=3, nNull=10,t.end=10
))
}

