% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvarss.R
\name{plotPost.tvarss}
\alias{plotPost.tvarss}
\title{Plot the posterior of a 'tvarss' parameter}
\usage{
plotPost.tvarss(x, varName = NULL, relative = TRUE, main, xlab, ylab)
}
\arguments{
\item{x}{a 'tvarss' or 'rjags' object}

\item{varName, }{NULL (default) or character indicating names of parameters in \code{x}. If NULL, plots all time-varying parameters}

\item{relative}{logical, if TRUE (default), at each time step the probability density of all values is divided by the maximum.}

\item{main, xlab, ylab}{character vector indicating the plotting labels to use. If a length-1 vector is supplied, the same character will be used for that label for all parameters}
}
\value{
invisibly returns a named list of (possibly relative) probability densities
}
\description{
Plot a heat map representing the probability density of a 'tvarss' parameter that varies over time
}

