% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_ac.sos.R
\name{roll_ac.sos}
\alias{roll_ac.sos}
\title{Rolling Autocorrelation, SOS}
\usage{
roll_ac.sos(X, window_elapsed, by = 1, n = 1, phase = 1, vars, lakes,
  subWindow = FALSE, subStat = FALSE, DETREND = FALSE)
}
\arguments{
\item{X}{a data.table of lake observations, or possibly list of data.tables. Should have columns named "variable", "lake", "y", and "x". See Details.}

\item{window_elapsed}{the number of observations that should have elapsed between the first and last observation; is the same as a "window size" when \code{n=1}}

\item{by}{integer indicating the number of time steps to increment the window as it "rolls". Can be a list or a vector if \code{X} is a list. Can affect rolling window or statistic subsample, see Details.}

\item{n}{integer indicating to subsample to every n-th observation when performing rolling window. Can be a list or a vector if \code{X} is a list. Can affect rolling window or statistic subsample, see Details.}

\item{phase}{starting datum to use when subsampling}

\item{vars}{a character, possibly vector, indicating which variables in "variable" column of \code{X} should be analyzed.}

\item{lakes}{a character, possibly vector, indicating which lakes in "lake" column of \code{X} should be analyzed.}

\item{subWindow}{logical, should the number of rolling windows by subset (uses the \code{by} and \code{n} arguments)}

\item{subStat}{logical, should the statistic be calculated for a subset of the windowed data?}

\item{DETREND}{logical, detrend or no? Pass to \code{\link{roll_ts}}. Also see \code{\link{detrendR}}.}
}
\value{
list or list of data.tables containing rolling window statistic
}
\description{
Calculate autocorrelation in a backwards-looking rolling window; intended for data set collect as part of the SoS Cascade project.
}
\details{
I need to fill this in. One thing to remember is that 28-day nominal window size is hard-coded into this function right now.
}

