% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_ts.R
\name{roll_ts}
\alias{roll_ts}
\title{Roll TS}
\usage{
roll_ts(y, width = 288, by = 1, FUN = mean, x, ...)
}
\arguments{
\item{y}{numeric vector representing time series}

\item{width}{integer indicating number of samples per window}

\item{by}{integer indicating the number of samples by which to increment the window forward between windowed subsamples}

\item{FUN}{function to be applied to each window}

\item{x}{optional vector by which the observations in y are ordered (such as date-times)}

\item{...}{additional arguments to be passed to \code{FUN}

The function performs a backward-looking rolling window on a numeric vector. Makes use of the \code{\link{embed}} function, which might be slow for long windows (many observations per window). Internally also makes use of \code{\link{sub_embed}}.}
}
\value{
a vector with each value being the result of applying \code{FUN} to a window.
}
\description{
Apply a rolling window statistic to a time series
}
\examples{
# example data
x <- 1:50
y <- cumsum(rnorm(50))
dt <- data.table(x, y)

# do x and y values separately
# x value is an ordering value, like time
# y is the system state, like chlorophyll
xa <- roll_ts(dt[,x], FUN=max, width=5)
ya <- roll_ts(dt[,y], width=5)

# do both x and y at same time
dta <- dt[,agg_ts(y=y, x=x, width=5)]

# plot two sets of results
par(mfrow=c(2,1))
plot(xa, ya)
dta[,plot(x,y)]
}

