% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundGrid.R
\name{roundGrid}
\alias{roundGrid}
\title{Round Grid}
\usage{
roundGrid(x, frac = 1)
}
\arguments{
\item{x}{numeric vector}

\item{frac}{numeric indicating fraction for rounding;}
}
\description{
Round values to nearest centered value, so as to put on a grid
}
\details{
For example, 0.5 means to put things on a 0.5 degree grid, or to round everythign to nearest 0.5, such that all values end in .25 or .75 (the center of intervals of 0.5)
}

