% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLakeP.R
\name{stabClass}
\alias{stabClass}
\title{Stability Classification}
\usage{
stabClass(I, pars, func = mDXM, Xvals = seq(0, 15, length.out = 15),
  Mvals = seq(0, 1000, length.out = 15))
}
\arguments{
\item{I}{numeric scalar representing P input; reasonable values might be between 0 and 2}

\item{pars}{a named numeric vector of optional additional parameter values to be passed to \code{func}}

\item{func}{function whose stability is to be classified; nominally this is \code{mDXM}, a wrapper for \code{\link{modelDeterministicXM}}; currently this function only works with the default b/c \code{\link{getRoot_df}} hasn't been set up to accept other functions as an argument}

\item{Xvals, Mvals}{numeric vector of X (water P) or M (sediment P) values over which to search for equilibria}
}
\value{
a data.frame with a row for each equilibrium point found, and columns for the X,M coordinates of those points, a character describing its stability classification, the trace (tr) and determinant (Delta) of the Jacobian matrix at that point, the 'discriminant' value (tr^2 - 4*delta), and the parameter values supplied through I and pars.
}
\description{
Find equilibria for a model system of lake eutrophication and classify their stability
}

