% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_embed.R
\name{sub_embed}
\alias{sub_embed}
\title{Embed and Subsample a Time Series}
\usage{
sub_embed(x, width = 1, n = 1, phase = 1)
}
\arguments{
\item{x}{numeric vector}

\item{width}{integer size of window/ embedding dimension}

\item{n}{integer, sample every nth element of \code{x}}

\item{phase}{integer, start subsampling sequency on the nth element of \code{x}}
}
\value{
matrix
}
\description{
Embed a time series like in \code{\link{embed}}, and then subsample the rows of that resulting matrix.
}

