% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_samp.R
\name{sub_samp}
\alias{sub_samp}
\title{Sub Sample}
\usage{
sub_samp(x, n, phase = 1)
}
\arguments{
\item{x}{a vector to be subsampled}

\item{n}{an integer indicating to subsample every n-th datum}

\item{phase}{an integer indicating which sample should be the first to be sampled}
}
\value{
a sub-sampled vector
}
\description{
Sub-sample a vector
}

