% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvarss.R
\name{summarize.tvarss}
\alias{summarize.tvarss}
\title{Summarize a 'tvarss' Posterior}
\usage{
summarize.tvarss(x, FUN = "mean")
}
\arguments{
\item{x}{a 'tvarss'  or 'rjags' object}

\item{FUN}{a function, or character string indicating a function, to be applied to posterior samples}
}
\value{
a data.table with columns for each parameter, and rows for indices of that parameter. Scalar parameters will be repeated.
}
\description{
Apply a summary statistic across the posterior samples of each parameter in a 'tvarss' object.
}
\note{
JAGS and \code{R2jags::jags()} accepts non-scalar parameter, such as vectors or matrices. In the context of time-varying autoregressive state space models, some parameters are going to vary over time. E.g., a parameter "Phi" might be an N-length vector containing parameters \code{phi[n=1], phi[n=2], ..., phi[n=N]}. As such, the output of this function will have 1 column for all "Phi", with each row representing a different phi[n]. The data.table will have N rows.
}

