% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_xreg.R
\name{trend_xreg}
\alias{trend_xreg}
\title{Trend X-Reg}
\usage{
trend_xreg(exp.order, y)
}
\arguments{
\item{exp.order}{integer, the order of the polynomial. 1 is a linear trend, 2 a quadratic, and so on}

\item{y}{a vector that is presumably the response variable, only used to determine length/ size/ number of rows in regressor matrix}
}
\value{
a matrix with column names indicating the order of the polynomial
}
\description{
A matrix of (polynomial) trend regressors (X) over the length of a response variable (Y)
}
\details{
no intercept column is included, and the output matrix is scaled. The standardization (scaling) occurs *after* the squaring, such that the "trend1" (linear) column squared does *not* equal the "trend2" column (this should be obvious because the "trend2" column will have negative values due to the standardization post-squaring.)
}
\seealso{
\code{\link{interaction_xreg}} \code{forecast::fourier}
}

