% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvarss.R
\name{tvarss_wrapper}
\alias{tvarss_wrapper}
\title{TVARSS Wrapper Function}
\usage{
tvarss_wrapper(x, det = FALSE, fit_arP = FALSE, niter = 3000,
  thinout = 500, ...)
}
\arguments{
\item{x}{an object conaining a list of data.tables to analyze. Each data.table should have a column named "variable", of which some elements need to be "X". When variable=="X", the a column 'y' of the data.table will be taken to be the time series to analyze. The column 'y' should be a "ts" object, with a specified \code{stats::frequency}}

\item{det}{logical, to detrend?}

\item{fit_arP}{logical, if TRUE, will fit an TVAR(p)SS model}

\item{niter}{number of iterations}

\item{thinout}{number of samples of the posterior to save in ouput (reduced from niter via 'thinning')}

\item{...}{additional arguments to be passed to \code{\link{tvarss}}; note that the arguments \code{niter=3E3}, \code{thinout=500}, \code{parallel}, \code{nP}, and \code{oType="tvarss"} are already provided either as arguments to this wrapper function, or specified directly within this function.}
}
\value{
a list of class "tvarss" objects; the names of the elements of this list will be the same as the names of \code{x}
}
\description{
A wrapper function for tvarss() that handles 1) applying tvarss to a list of data.tables/ matrices containing time series to be analyzed; 2) detrending; 3) finding 'p' for AR(p) models; 4) calculating and extracting eigenvalues of the AR(p) models (and doing so in a way that handles them just like another 'parameter' from the model)
}

