% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLakeP.R
\name{mDXM_jac}
\alias{mDXM_jac}
\title{Wrapper for model when finding Jacobian matrix}
\usage{
mDXM_jac(t, state, pars, ...)
}
\arguments{
\item{t}{time, for compatibility with rootSolve::jacobian.full, but not used otherwise}

\item{state}{state variables as in \code{\link{modelDeterministicXM}}}

\item{pars}{paramters as in \code{\link{modelDeterministicXM}}}

\item{...}{additional arguments to pass to \code{\link{modelDeterministicXM}}}
}
\value{
rate of change of state variables
}
\description{
Wrapper for model when finding Jacobian matrix using \code{rootSolve} package
}

