% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescales-package.R
\docType{package}
\name{timescales-package}
\alias{timescales}
\alias{timescales-package}
\title{timescales: Nested Timeframes and Calendars for Modeling}
\description{
Discrete time representations for optimization and simulation models,
including calendar-based nested sets of timeframes, conversion between
alternative temporal resolutions, and mapping of data onto calendar grids
for modeling and visualization.
}
\details{
Successor to the 'timeslices' package; companion to 'geoscales' for the
spatial dimension.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/optimal2050/timescales}
  \item Report bugs at \url{https://github.com/optimal2050/timescales/issues}
}

}
\author{
\strong{Maintainer}: Oleg Lugovoy \email{olugovoy@optimalsolution.dev} (\href{https://orcid.org/0000-0001-5546-9875}{ORCID})

Authors:
\itemize{
  \item Oleg Lugovoy \email{olugovoy@optimalsolution.dev} (\href{https://orcid.org/0000-0001-5546-9875}{ORCID})
}

}
\keyword{internal}
