% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvarss.R
\name{tvarss}
\alias{tvarss}
\title{TVARSS}
\usage{
tvarss(Y, nP = 1, niter = 1000, tvMean = FALSE, parallel = FALSE,
  oType = c("jags", "tvarss"))
}
\arguments{
\item{Y}{vector of observed values of time series}

\item{nP}{scalar, integer; order of the AR process}

\item{niter}{scalar, integer; number of MCMC iterations}

\item{tvMean}{logical; if TRUE, include a time-varying 'constant' parameter indicating that the mean of the time series changes over time}

\item{parallel}{logical; use parallel computing?}

\item{oType}{character indicating type of output; 'jags' is default, outputs an 'rjags' class object. If 'tvarss', output type is class 'tvarss', which is just a subset of the original 'rjags' object (just includes posterior samples of parameters as a list)

The model is as follows.
Observation:
\eqn{y_t = z_t + v_t}
Process:
\eqn{z_t = C_t + \Phi_t*B^p(Z_t - C_t) + w_t}
\eqn{B^p} is backshift of order p; if \eqn{p=2}
\eqn{B^pZ = [z_{t-1}, z_{t-2}]^{-1}} ; note that the [ are used here to indicate a matrix
Parameters are time-varying:
\eqn{\Phi_t = \Phi_{t-1} + \epsilon_t}
\eqn{C_t = C_{t-1} + \xi_t}}
}
\value{
a model of class 'rjags' or 'tvarss'
}
\description{
Fit a time-varying autoregressive state-space model in a Bayesian framework using JAGS
}

